Fix TDX PAMT int type out of range issue for large memory size system
On large memory size system (eg. 1T), the `int` type used by
`pamt_entry_nr` would be out of range after multiply it by PAMT entry
size, change the type to `unsigned long` to fix this issue.

+++ arch/x86/virt/vmx/tdx/tdx.c
@@ -1056,7 +1056,7 @@ static unsigned long tdmr_get_pamt_sz(struct tdmr_info *tdmr,
 				      enum tdx_page_sz pgsz)
 {
 	unsigned long pamt_sz;
-	int pamt_entry_nr;
+	unsigned long pamt_entry_nr;
 
 	switch (pgsz) {
 	case TDX_PG_4K:
